[DM-23923] Fix spawn issue
Issue with following the redirection, so I'm being more explicit
about checking the lab URL.  This is a bit tricky since it actually
sends me to a spawn-pending page, which is a little confusing to
handle automatically.  It will only do the redirect once the spawn
is finished.  Also when comparing the URL, make sure to compare it
as a string.

Note, if you don't follow redirects to the lab, you'll have to go
through all the redirects for an oauth flow back to the lab.  This
gives you an additional cookie that you use to auth to the lab.