[DM-25328] Fancy up the logging
Since there can be a lot of different streams, break up the logging
a bit.  I decided that I wanted more information through the rest
api, so I'm using the logging API to make one logger per user, and
then apply the structlog magic to it to get it to print out all nice
and to the console.  Then this file (which contains only that user)
is available via the restful api.  I'll be adding more status stuff
later on too.