[DM-25653] Fix exception while deleting monkey
There was a cancelled and timeout exception thrown when closing
down a job.  So we have to catch those to make sure that we clean
up properly.  This isn't the best, but it reduces a lot of complexity
from the business logic (lol) as to where to stop and how long it
takes to stop.  At least this happens right away.