[DM-26303] Allow spawning of lab to give up
Sometimes the hub and mobu seem to get confused about where we are
in the spawning, and mobu doesn't have a great way to get notified
if there's an error - we just keep retrying.  But after 900 seconds,
the hub will give up, so mobu should try for about that amount of time,
then give up.