[DM-29321] Bail out on bad progress URL status codes
Before I would wait since it seemed like we would get a bunch of 200
status codes.  But if the first status code is a 500, apparently we
swallow it and keep waiting.  The next request gets a 200, even though
it isn't waiting at all for progress.  This is why it takes a whole
timeout of running through that loop, but really we should signal the
first error we see.  This won't fix anything, I don't think, but it
should reduce some noise and change some timings.