Convert to FastAPI
Convert to a FastAPI web application instead of aiohttp.  For now,
leave all of the internals the same, including using aiohttp to
make outbound requests.

Add a mock JupyterHub (currently a simple state machine) and move
the mock Gafaelfawr into a test module.  Use those to add a test
case that starts a login loop, retrieves its status, and deletes it.

Currently, this does not use pydantic models for the API or provide
any meaningful API documentation.  This will change in future
revisions.