Support multiple users in JupyterHub mock
Rather than configuring the mock with a single user, store state
for multiple users and extract the user from the token presented
to the hub.  Encode the username in the fake Gafaelfawr token.

Unfortunately, aioresponses cannot see headers set on the session
(see https://github.com/pnuckowski/aioresponses/issues/111), so we
have to add a wrapper around the ClientSession that adds the
Authorization header to each call.