Refactor business classes
Remove the spurious use of dataclasses for objects that are not
data transfer objects.  Unify Business and BusinessTime since
there's no need for two layers of inheritance.  Move construction
of clients into the __init__ function of the business classes and
remove unnecessary duplicate code from child classes.

Restructure how business is created and assigned to a monkey to
have the monkey create the business directly from a type that's
passed in.  This makes it easier for the monkey to pass some of its
internal state to the business, which in turn avoids circular
references between the monkey and the business.