Refactor ClientSession handling
Share a single ClientSession for monkey alerting and token
creation across all monkeys.  Continue to create a separate
ClientSession for each JupyterClient, since it has some custom
configuration, but close that ClientSession when the Business
is stopped.  Create and close the shared ClientSession when the
MonkeyBusinessManager is initialized and cleaned up.

This fixes the remaining errors in testing about unclosed sessions.