Further refactor aioresponses mocking
The mock JupyterHub was still only registering aioresponses
callbacks for a specific user.  Actually make this generic and
handle any user, and confirm that the user specified in the URL
matches the user in the token.

Move aioresponses into a fixture to remove some unnecessary
indentation.

Turn admin_token into an autouse fixture so that it doesn't have
to be specified for each test that may create users.