Refactor business classes
Break up the huge run methods into lots of smaller methods, each
of which tries to do one timed event.  Use this to share more
code with subclasses.

Move more of the setup outside of the try block and scope the try
block to only the code that should count towards successes and
failures.