Refactor handling of timings
Turn Stopwatch into a context manager that stops itself when the
context is exited.  Create a Timings class that wraps the list of
Stopwatch instances and handles creating new ones with the proper
previous stopwatch parameter.

Redo the timers so that the minimum interesting code is inside
the range of the stopwatch.