Add test for JupyterPythonLoop
This requires adding support for creating and deleting a kernel
to the mock JupyterHub.  Unfortunately, we can't mock ws_connect,
so we mock out the code execution logic entirely.  That will have
to be tested on a live cluster.

Running two tests revealed that we didn't clear the state of the
MonkeyBusinessManager between tests.  Clean that up during shutdown
of the application.

Do not attempt to stop a monkey that is already stopped, since we
have the state available to us.