Clean up running code in NotebookRunner
Change self.code to self.running_code to match its semantics and
clear it as soon as we stop running the code.  The primary purpose
of including this in the dump is so that we can see if a cell is
taking an unexpectedly long time, so we only want the code if that
code execution is currently in progress.