Switch to Pydantic models
Accept and return Pydantic models instead of dicts.  Plumb this
down to all the components, so that they take models as
configuration rather than parameters or dicts.  Return models from
dump() rather than dicts.

Return the name of the monkey as part of its metadata.

Return the newly-created monkey from the POST to /user instead of
only the monkey name, and return a 201 status with a correct
Location header.