Use better typing for configuration
Use None instead of the string "None" for configuration options
that aren't set, except for the environment URL (where I don't
want to have to check against None everywhere it's used).  For it,
default to the empty string, which will clearly break everything
if it's not set.  Add test support for setting it to a test string,
which will produce clearer test error messages than the previous
default of nublado.lsst.codes.

Add awaits to two tests to prevent busy-looping.