Rewrite monkeyflocker to use the new flock API
Most of the logic in monkeyflocker can be removed because the new
flock API achieves the same ends.  Turn it into a simpler wrapper
around the flock API that takes a YAML configuration file for a
flock of monkeys to start, or the name of a flock to gather reports
from or stop.

Add a proper test for monkeyflocker that actually talks to mobu.