Poll the Hub URL instead of pending spawn URL
After spawning a lab, the web browser retrieves a progress URL.
It doesn't keep trying to reload the spawn-pending URL, and we're
seeing occasional 503 errors polling that URL.  My theory is that
the Hub decommissions the spawn-pending URL after the lab is
spawned and we're requesting it after that has happened.

For now, rather than trying to understand the results of the
progress URL, poll the hub URL instead.  Once there is a spawned
pod, it will redirect to the lab URL.  See if this is more
reliable.