Fix desynchronization in web socket
We return immediately when we get the first stream reply from an
execute request on the web socket.  Now that we're reusing the
web socket, this results in only the first execution result being
returned.  That's because we leave the execute_reply message in
the web socket and don't check its message ID.  The execution
sequence is therefore:

- Send the first reqeust
- Retrieve a stream response and return it
- Send the second request
- Retrieve the execute_reply response from the first and return
  the empty string
- Send the third request
- Retrieve the stream response from the second request and ignore
  it because the message ID is wrong
- Retrieve the execute_reply from the second response and return
  the empty string

This continues until the session is closed.

Fix this by, one, always checking the message ID of the response
and complaining about unexpected responses, and two, accumulating
all stream responses in a result buffer and not returning them
until we see the execute_reply.