Wait for the lab to shut down, fix shutdown
Check whether the lab is still running after shutdown and wait up
to ten seconds for the lab to shut down completely.  This will
help avoid occasional spurious issues when we delete the lab and
then start using it before it shuts down.

This change uncovered various shutdown concurrency isseus.  Fix
the state handling inside the monkey so that the wait after an
error when respawning the business doesn't block shutdown.  Ensure
that we acknowledge the business shutdown command even if we error
in startup after it was sent.  Shut down all flocks in parallel
during global shutdown.