Redo how lab status detection is done
We are having continual problems analyzing redirects from the hub
URL to determine the status of the lab, and get unexpected results
if a spawn is currently pending.  Try a different approach by
separating checking whether a lab is fully stopped from checking
if a spawned lab has started.

To check if a lab has fully stopped, go to hub/home and see if
there is a Start My Server button.  This appears to happen only if
the lab is fully stopped (not shutting down, not spawning, not
running).

To check if a spawned lab has finished spawning, attempt to go
directly to the lab.  This will redirect to either a spawn pending
page or to a 503 error if the lab is not fully running.

Also clean up some typing for the web socket connection and move
the lab login before session delete into the business so that we
have better logging and timing.  Rearrange the logging and remove
some unnecessary logging (although there's more of this to do).