Improve Slack alerting
Send more elaborately formatted Slack alerts for unexpected
Jupyter responses, code execution failures, and timeouts spawning
the lab.  Only the former is currently tested by the test suite.

This is implemented via a new class hierarchy of rich exceptions
that store a bunch of additional information and support a method
to generate a Slack message containing the details of the exception.

Now log the body of the error message returned by Jupyter in the
event of an unexpected status.