Use progress and user API to get lab status
After a spawn, use the progress EventStream API to wait for the
lab to spawn and to log any messages returned during the spawning
process.  This required some restructuring to accomodate an async
iterator over spawn progress results.  As a side effect, this fixes
problems with stopping business during the wait for a lab to spawn,
which previously would be delayed to the end of the wait interval.

Add a configuration option for how long to wait for a lab to spawn,
since this was needed for the test suite.