Move the wait for deletion into the business
Rather than waiting for a lab to delete in the JupyterClient,
move that logic into the JupyterLoginLoop business so that it can
use the pause method and interrupt the wait if signaled to shut
down.  Increase the default timeout for lab deletion to 30 seconds,
make it configurable, and report an error to Slack if the lab
doesn't go away by the end of the timeout.