Include the node in code execution errors
Do preliminary code setup when creating the session instead of when
executing code so that JupyterPythonLoop and NotebookRunner can more
easily share code.  Get the node on which the notebook is running
and add it to the timer annotations, and from there include it in the
exception and Slack alert.

Stop adding the code and result as an annotation on timers.  This
bloats the timer data structure excessively and duplicates information
found in the log, as well as adding unnecessary code complexity.

Rename timer annotations to use the plural.