Stop using an iterator to read notebooks
NotebookRunner was being excessively fancy in how it scanned for
notebooks, which meant it would go into an infinite loop if given
a directory devoid of notebooks.  Instead, read in all the
notebooks in a directory via the Path object, raise an exception if
none are found, and then use pop to exhaust that list.

Randomize the order in which we run the notebooks each time through
the loop so that we get better test coverage even if one notebook
is failing.