Clean up event names for NotebookRunner
Move all of the ancillary information about cell execution into
timer annotations.  Use annotations to pass the notebook name to
the exception (via the Stopwatch), which eliminates a try/except
block in NotebookRunner.  Add the cell ID to the annotations and
to the corresponding alert formatting.

All event names are now static in preparation for coalescing
timer data.  Per-configuration variations are now captured entirely
in annotations.