Disable propagation of monkey logs
Something underneath pyvo.dal.TAPService adds a root logger during
the constructor and whenever search is called, which was duplicating
every log message on standard output.  I was unable to find what bit
of code was responsible, but disabling log message propagation (which
we were not using anyway) avoids the problem.