Stop recreating the business on failure
To keep accurate long-term statistics, we cannot recreate the
monkey business since then we lose all of the history.  Fixing this
required moving the pause after error handling into the Business
class and ensuring that it correctly handles the case where the
monkeys are being shut down while in an error state.

As a bonus, this simplifies the logic in Monkey somewhat.