Allow web socket messages of any size
Since the JupyterLab interface doesn't impose a message size limit, it makes sense to treat the notebooks run in CI the same way.  We may want to do a sweep occasionally to see how big the messages are, but I'm not sure it's the right thing to fail on them.