Increase expiration of tokens to one year
nublado2 is now better at checking the expiration of tokens and
attempting to refresh them, so ensure the token is valid for
longer than we're likely to leave mobu running.

A more elegant and correct approach would be to create a new token
when the old one is about to expire, but it's also much more
complex given mobu's structure, so defer that for now since it
may be handled automatically in the future via noteburst.