Add GID support
Allow explicitly specifying the GID, either per user or with a
spec.  Default to setting the GID to the same as the UID, which was
the behavior previously forced in nublado2, if the UID was specified.
If it was not specified, leave the behavior to Gafaelfawr, but
document that synthetic user private groups need to be enabled.