Convert configuration to Settings
Use the Pydantic Settings class for configuration instead of a
dataclass with os.getenv initializers. Fix the typing of URLs to
validate syntax, which unfortunately requires allowing None for
the test suite and adding more verification elsewhere. Be a bit
more robust in URL construction to allow for trailing slashes in
the configured URLs.

Switching the type to HttpUrl requires doing some weird things to
override the defaults in fixtures, but this is apparently the
correct Pydantic way of doing it.