Remove types from docstrings
Take advantage of Sphinx's new support for type annotations and
remove type information from docstrings except for the return value,
which now doesn't need a variable name.

Remove docstrings from routes, and remove __all__ in the handler
module since we won't generate documentation for that (even if we
get around to generating documentation).