Redo monkey logging and state machine
If the profile is production, don't log all the monkey output to
the main application log. Keep it only in the per-monkey log. This
generates huge amounts of logging volume that we essentially never
look at, and those logs aren't in the production log format.

Maintain two separate loggers in the monkey code and log some events
to the global logger so that we get a bit of status even though we're
not showing the output of every monkey.

Tweak the state machine to hopefully be clearer about shutting down
a monkey due to errors, and to avoid calling close on the business
more than once.