Add notebook to Slack error reports
The notebook was included in the main part of the Slack message
when reporting a code execution failure, but not when reporting
other problems such as web socket issues. Add the notebook to the
cell in the common fields added from annotations so that it will
always be present even for exceptions that don't add it to the
main Slack message.