Tidy up the Business implementations
Reorganize the methods on Business to make it clear which ones
should be overridden, which are the external API, and which are
helper functions for subclasses. Fix more places where we can
return a boolean to indicate whether the business should be
stopping, and improve the logging in some of the business classes.