Restructure the Business class hierarchy
Rather than mushing all the business options together in one blob
and ignoring some of them depending on the business, use separate
classes and Pydantic type validation to ensure that only the correct
options for the specified business are given. This has a lot of
cascading changes on the type system elsewhere, including how a
business is instantiated by a Monkey and how monkey and flock status
are reported.

To get the typing right, turn some of the business classes into
generic classes parameterized by the type of their options.