Merge JupyterLoginLoop and JupyterPythonLoop
We never just create a lab without executing some code on it, and
there doesn't seem to be a reason why we would ever want to do that.
The distinction between JupyterLoginLoop and JupyterPythonLoop was
therefore unnecessary and they could be combined.

Use this as an opportunity to refactor code execution in a Nublado
lab into a separate abstract base class which can be used by both
JupyterPythonLoop and NotebookRunner without confusing the logic of
the two. This eliminates some runtime checks and other awkwardness.