Handle failures during TAPQueryRunner setup
TAPQueryRunner previously was creating a TAP client in __init__,
which meant that failure to reach the TAP server caused the entire
flock creation to fail. Move creation of the client to startup to
correctly follow the Business API, and wrap it in a timer and better
exception handling so that we get nice Slack reports.

A business wasn't counting failures during startup as a failed
execution, which meant that monkeys that kept failing during startup
would show as 100% successful in the summary. Increment the failure
count if startup raises an exception to avoid this.