Refactor mobu state management
In preparation for adding new functionality to run a single
business once to completion and report the results, refactor mobu
to use a ProcessContext and RequestContext similar to Gafaelfawr
and the new lab controller. This provides a hook to create a
per-request Factory, which will be needed by subsequent changes.

Mostly this is just moving code around, but add documentation for
the FlockManager class and use the proper FastAPI function for
getting the URL of a resource.