Add support for running a business once
Sometimes rather than running a business continuously in the
background and reporting the results to Slack, we want to run that
same business once to completion and then report the results,
particularly whether it succeeded or failed. We'll use this later
to do things like check whether a new deployment of Phalanx is
functional.

Add support for this to mobu via a POST /run route and a new concept
called a Solitary. This is parallel to a Flock, except rather than
managing a possibly large number of monkeys doing business in the
background, it manages a single monkey doing its business in the
foreground until it finishes.