Add support for the new Nublado lab controller
Change the way that the image is specified to businesses that
inherit from NubladoBusiness to put more of the model into the
business model and rename it to Nublado to be more consistent about
branding. (This is not generic JupyterHub support.) Add support for
the various ways of specifying an image that are supported by the
new lab controller.

Make cachemachine support conditional, and if disabled, use the new
direct POST protocol for spawning an image by class, reference, or
tag that's spelled out in SQR-066.

This means that the running business no longer knows the exact lab
that was spawned, since we previously relied on cachemachine to tell
us that and in the new setup there's no need to ask the lab controller
anything directly. Discover the lab image instead by running some
Python code in it after it was spawned, similar to how we discover the
node name. This will get the description under either the old or new
spawning method, and the full reference when the lab was spawned by the
lab controller.