Make the error idle time configurable
Add a configuration option to set the error idle time, and
increase the default for Nublado businesses to the ten minute
timeout imposed by JupyterHub when it's waiting for the lab to
spawn. It won't let us try to delete the lab before then anyway,
so attempting it just results in alert spam.