Be more robust when getting the running image
Stripping before splitting means that we raise a random ValueError
if one of the environment variables was missing. More generally,
we got several exceptions that I think may be due to the lab getting
shut down while we were asking it questions. Be more robust when
parsing the response and treat any unrecognized response as providing
no data.