Revert error_idle_time change from 5.0.0
In mobu 5.0.0, I had changed the default error_idle_time for those
businesses that use Jupyter labs due to problems with recreating
the lab after an error. This turned out to be a bug in the new lab
controller which should now be fixed, so revert this back to 60
seconds.