Format node and cell as blocks
In Google Kubernetes Engine, nodes are often longer than the useful
width of a field. Cell is also often longer than the width of a field
when there is notebook information and a UUID available. Format both
as full Slack Block Kit blocks instead of fields so that they get the
full width of the message and don't mess up the column formatting.