Convert to httpx
Follow our other packages in using httpx as the HTTP client library
instead of aiohttp. Remove all of the remnants of aiohttp support
and make use of SlackWebException from Safir now that the HTTP client
exceptions come from httpx. Switch to httpx-ws for WebSocket support
and httpx-sse for EventStream events.

As part of the reworking of the WebSocket layer, manage lab sessions
with a context manager. This is a better match for how httpx-ws thinks
about WebSocket sockets.

Replace the mock layer that was using aioresponses with respx.