Split out Gafaelfawr storage layer and use models
Rather than embedding the Gafaelfawr client in the AuthenticatedUser
model, separate out a Gafaelfawr storage client. Use copies of the
Gafaelfawr models rather than sending and receiving raw JSON. These
can be replaced with the real Gafaelfawr models once those are
available from PyPI.

Take advantage of this to use Pydantic for the parsing and capture
Gafaelfawr parsing errors in a Slack-reportable exception.