Remove WebSocket message size limit
Add a constant so that we can configure a WebSocket message size
limit in the future, but remove the default limit to match the
previous aiohttp behavior. We have some notebooks that return images
as part of cell execution and they exceeded the default 1MB limit
on WebSocket message sizes.