Stream monkey logs instead of using FileResponse
Starlette FileResponse no longer works to return monkey logs
because it now commits to a Content-Length when starting to stream
the file but doesn't stop sending data at that cutoff, so it sends
more data than Content-Length since the log file is changing, and
that in turn causes an HTTP protocol failure that prevents Chrome
(at least) from successfully downloading the file.

Instead, simulate that with StreamResponse while setting a
Content-Disposition header.