Configure Ruff and fix things it found
Ruff is an interesting new (and extremely fast) Python linter that
replaces isort, flake8, and a whole bunch of plugins. Configure it
with fairly aggressive settings but a lot of exclusions, and fix
all the problems it found.

Most of these changes are converting Optional[type] to type | None.
I previously was sticking with Optional in some specific cases, but
given that Ruff can autofix all other uses of Union and the
distinction I was drawing was very nitpicky, take this approach
instead.