Move cachemachine config to business options
Allow use of cachemachine and the cachemachine image policy are
now configurable in the business options rather than globally.

Move the creation of the cachemachine client into the business
layer to make it easier to move JupyterClient into Safir and
simplify its constructor. Move annotation of cachemachine exceptions
with username information into JupyterClient to simplify the
cachemachine client.