Generate less data when summarizing flocks
Previously, to generate a flock summary, mobu first converted all
information about the running monkeys to output data structures and
then added up success and failure counts from those. Because of the
huge amount of timing data that mobu gathers, this took long enough
(and is not async, so it blocks everything else) that other
in-progress calls timed out. It also meant that the summary end
point appeared to hang.

Avoid this by reading the success and failure counts directly from
the running monkey business, avoiding the need to dump and thus
convert their data. This is a temporary fix until the way that
timing data is stored can be overhauled.