Fix XSRF cookie and header handling for JupyterLab
Previously, mobu created a fake XSRF cookie and header and sent them
blindly to JupyterLab. JupyterLab now expects the cookie to be in a
specific format and is rejecting our fake one.

Instead, properly implement the same protocol as the JavaScript client:
retrieve the _xsrf cookie from the response from the lab and lift it
into the X-XSRFToken header. Raise an exception if the lab does not
send an _xsrf cookie so that we'll catch further changes more quickly.

This is complicated by the fact that the hub and lab use separate XSRF
tokens restricted to different paths, so we need to track both tokens
and send the appropriate header based on which API we're talking to.