DM-48101: Sentry for notifications and timings
* Replace Safir `SlackException` functionality with Sentry error
reporting.
* Replace `timings` functionality with Sentry tracing.

The biggest conceptual change is that instead of putting error details
on the exception objects themselves, they get put into the
global-per-business-execution Sentry scope, and whatever is in the scope
is sent to Sentry when an error occurs (except for a very few cases
where we still attach info directly to exceptions).

This instruments custom scopes and transactions and fingerprints, which
are described in the new `sentry.rst` file in the docs.

Finally, this uses some more general Sentry helpers proposed for Safir
in https://github.com/lsst-sqre/safir/pull/366.