Use HumanTimedelta for all business durations
Switch to the Safir `HumanTimedelta` Pydantic type for all business
durations, which supports a superset of the functionality of the
previous mobu-specific type. This allows any time duration in a
business configuration to use human-readable strings like `60m`, but
still supports a simple number of seconds.

Delete the now-unused `StopwatchData` class.