Treat the Slack alert webhook as a secret
The Slack alert webhook URL is an incoming webhook that allows
direct Slack posting as mobu, so it is a secret and should be treated
as such. Change its type in the mobu configuration and pass it
directly to the Safir Slack library, since those classes support
`SecretStr` as an input type.