Update GitHub Actions and tox dependencies
Separate the tox dependencies into a separate pinned requirements file
that is also updated by `make update-deps`. Resynchronize the GitHub
Actions configuration with the current templates, which stops using
neophile, uses the tox requirements file, and adds support for merge
queues (although we're not yet using that). Fix the condition for Slack
reporting of failed periodic checks and add a periodic check for the
documentation build as well.