DM-49320: Refactor notebook businesses
The `NotebookRunner` business currently can accept two kinds of options,
one which specifies `max_executions` and runs a constant number of
notebooks before refreshing the lab, and one which specifies
`notebooks_to_run`, which takes a list of notebooks and runs them all
before refreshing.

To keep consistency with the overall Mobu design, there should be a 1:1
mapping between businesses and business option types. We're about to add
a third variation on behavior (don't ever refresh a lab), so by the rule
of three, let's clean up this debt and do this refactoring now.

Another piece of debt clean up done here is to split the notebook
filtering logic into a helper class.