Skip summary message if there are no flocks
Stop sending summary messages if there are no flocks running, which
will reduce the alert spam on T&S status channels where we currently
don't use mobu. The original thought was that this would serve as a
sort of heartbeat and would alert us if all flocks were stopped, but
we mostly don't manually stop flocks and the annoyance of the empty
messages probably exceeds the benefit as a heartbeat.