DM-49361: Multi-instance mode
Enable a multi-instance mode by setting `replicaCount` to something
greater than 1. In this mode, the number of monkeys specified in a
flock's `count` will be spread evenly across the number of replicas. It
is assumed that multiple instances of mobu will be running with the same
config, so:

* The web API will only return info from a single
  replica. Getting info from every instance will require forwarding
  requests to individual pods.
* The GitHub refresh integration will not work because only one pod will
  get the webhook from GitHub.

Multiple instances are only meant to be used during load testing, when
more than 1 CPU is needed, and folks are aware of the limitations. If we need to run our normal flocks with more than 1 CPU, a more robust solution will be needed.

Also fixes a bug where `execution_idle_time` was not waited at all for
empty notebooks.