Port mobu to the new Nublado client
Adjust mobu's Nublado integration for the new Nublado client, which
uses Repertoire for service discovery. The mock API was entirely
reworked, so this requires several changes to the test suite.

Stop re-publishing the Nublado error attachment in the Python loop
code, since the Nublado client now does this properly for Sentry
under a slightly different name.